[*] change long double to double

+++ core/include/dap_common.h
@@ -604,8 +604,8 @@ extern "C" {
     ({                                                          \
         __typeof__(a) _a = (a); __typeof__(b) _b = (b);         \
         if (!( \
-            (((long double)_b > 0 && (long double)_a > (long double)dap_maxval(_a) - (long double)_b) || \
-            ((long double)_b < 0 && (long double)_a < (long double)dap_minval(_a) - (long double)_b)) \
+            (((double)_b > 0 && (double)_a > (double)dap_maxval(_a) - (double)_b) || \
+            ((double)_b < 0 && (double)_a < (double)dap_minval(_a) - (double)_b)) \
         )) { (_a += _b); } \
         (_a); \
     })
@@ -614,8 +614,8 @@ extern "C" {
     ({                                                          \
         __typeof__(a) _a = (a); __typeof__(b) _b = (b);         \
         if (!( \
-            ((long double)_b < 0 && (long double)_a > (long double)dap_maxval(_a) + (long double)_b) || \
-            ((long double)_b > 0 && (long double)_a < (long double)dap_minval(_a) + (long double)_b) \
+            ((double)_b < 0 && (double)_a > (double)dap_maxval(_a) + (double)_b) || \
+            ((double)_b > 0 && (double)_a < (double)dap_minval(_a) + (double)_b) \
         )) { (_a -= _b); } \
         (_a); \
     })
@@ -625,14 +625,14 @@ extern "C" {
         __typeof__(a) _a = (a); __typeof__(b) _b = (b); \
         if (!( \
             /*_a positive*/\
-            ((long double)_a > 0 && ( \
-                ((long double)_b > 0 && (long double)_a > (long double)((long double)dap_maxval(_a) / (long double)_b)) || \
-                ((long double)_b < 0 && ((long double)_b < (long double)((long double)dap_minval(_a) / (long double)_a))))\
+            ((double)_a > 0 && ( \
+                ((double)_b > 0 && (double)_a > (double)((double)dap_maxval(_a) / (double)_b)) || \
+                ((double)_b < 0 && ((double)_b < (double)((double)dap_minval(_a) / (double)_a))))\
             ) || \
             /*_a negative*/\
             (_a <= 0 && ( \
-                ((long double)_b > 0 && (long double)_a < (long double)((long double)dap_minval(_a) / (long double)_b)) || \
-                (_a != 0 && (long double)_b < 0 && (long double)_b < (long double)((long double)dap_maxval(_a) / (long double)_a))) \
+                ((double)_b > 0 && (double)_a < (double)((double)dap_minval(_a) / (double)_b)) || \
+                (_a != 0 && (double)_b < 0 && (double)_b < (double)((double)dap_maxval(_a) / (double)_a))) \
             ) \
         )) { (_a *= _b); } \
         _a; \
